Added DownloadAssetsServiceWorker and Fragment to download steps

+++ app/src/main/java/com/videostreamtest/ui/phone/downloads/fragments/DownloadFlagsFragment.java
@@ -115,7 +115,7 @@ public class DownloadFlagsFragment extends Fragment {
         downloadsViewModel.getCurrentDownloadTypeInformation("flags").observe(getViewLifecycleOwner(), generalDownloadTracker -> {
             if (generalDownloadTracker != null) {
                 descriptionView.setText(getString(R.string.download_general_message)+generalDownloadTracker.getDownloadCurrentFile());
-                downloadProgressbar.setMax(generalDownloadTracker.getDownloadTypeTotal());
+                downloadProgressbar.setMax(Math.toIntExact(generalDownloadTracker.getDownloadTypeTotal()));
                 downloadProgressbar.setProgress(generalDownloadTracker.getDownloadTypeCurrent());
                 if (generalDownloadTracker.getDownloadCurrentFile().equalsIgnoreCase("done")) {
                     nextButton.setVisibility(View.VISIBLE);
