+++ app/src/main/java/com/videostreamtest/ui/phone/downloads/fragments/DownloadRoutepartsFragment.java
@@ -113,7 +113,7 @@ public class DownloadRoutepartsFragment extends Fragment {
         downloadsViewModel.getCurrentDownloadTypeInformation("movie-routepart-images").observe(getViewLifecycleOwner(), generalDownloadTracker -> {
             if (generalDownloadTracker != null) {
                 descriptionView.setText(getString(R.string.download_general_message)+generalDownloadTracker.getDownloadCurrentFile());
-                downloadProgressbar.setMax(generalDownloadTracker.getDownloadTypeTotal());
+                downloadProgressbar.setMax(Math.toIntExact(generalDownloadTracker.getDownloadTypeTotal()));
                 downloadProgressbar.setProgress(generalDownloadTracker.getDownloadTypeCurrent());
                 if (generalDownloadTracker.getDownloadCurrentFile().equalsIgnoreCase("done")) {
                     nextButton.setVisibility(View.VISIBLE);
